Use 'get_hash' instead of 'id' when comparing UpfData in test.
In the 'immigrate' tests, the UpfData for Si was compared by id.
When using a AIIDA_TEST_PROFILE, this led to errors because a
different UpfData of the same kind could already be present, and
the immigration did not necessarily pick the one created for the
test.
To make the test independent of the UpfData PK, they are now
compared by hash.