Use `get_hash` instead of `id` when comparing `UpfData` in immigrate test (#527)
The point of the immigrator is that it correctly reconstructs the pseudo
potential nodes from the input files. To properly test this, we should
then of course test that the content is identical to the original, not
that exactly the same node instance is returned, if it weren't just for
the fact that in the real situation, the original node might not even
exist yet.