`PpParser`: fix whitespace bug in the `parse_gnuplot2D` method (#534)
The code was assuming that a line either contained valid data or was
completely empty, in which case it is skipped. However, sometimes the
code can produce empty lines with whitespace, such as spaces, which
would crash the code. Stripping the line before further processing
solves the problem.

Since this bug relies on whitespace in the test data file, it was not
being caught, because the pre-commit would strip trailing whitespace.
That is why we reintroduce the whitespace manually and add any files
with a file suffix that ends in `out` in the `tests` folder to the
exclude list so it is not automatically removed on commit.