`PpParser`: improve the Gaussian cube file parsing algorithm. (#535)
The current implementation assumes that the number length is 13 digits,
which is not transferable. In addition, it uses nested loops, which are
slow. The algorithm is changed to replace the nested loops by chomping
the float numbers one-by-one and reshaping the final one-dimensional
list into the three-dimensional matrix that is expected.