Small change to test_pw.py.
The second part of the statement `orm.Dict(dict=settings) or orm.Dict()`
is never executed since `orm.Dict(dict=settings)` can't be a falsy
object. Moreover, if `settings` is equal to `None` or `{}`, the
`orm.Dict(dict=settings)` is exactly the same as `orm.Dict()`. Hence
`or orm.Dict()` can be safely removed.