`PwBaseWorkChain`: do no assume `output_band` exists in band sanity check
The `output_band` output of the `PwCalculation` is optional and, for
example, when the setting `no_bands=True`, the band files will not be
retrieved and consequentially not parsed. The process handler checking
the bands `sanity_check_insufficient_bands`, however, assumed the output
node is always there. This is changed to simply skip the check if the
node is not there.

Note that we do not check in the workchain whether the output should
have been there, because the sanity check is only triggered for
successful calculations, and if the output should have been there, the
calculation should have been failed. If this is not the case, that is a
bug in the parser and should be addressed there.