Fix parsing problem with van der Waals functionals+stress
Previously, the flag `vdw_correction` was set in the raw PW parser
only if the string 'Carrying out vdW-DF run using the following parameters:'
was found.
However, `vdW-DF is not the only vdW functional, there are many more.
Therefore, e.g. in my case using a rVV10 functional, the string was more like
'Carrying out RVV10 run...', and the capitalization might also differ.

Since this flag is used later to parse lines having the text
'Non-local correlation energy' in it, I now change the setting of the flag
to the existence of this string.

As a consequence, the actual manifestation of this bug was that
(beside the key `energy_vdw` missing in the output) that a generic
'Error while parsing stress tensor.' was raised.
After debugging the cause was that tsince the vdw flag was not
set, the search for `P=` was failing as it was searching not far
enough (was: `for k in range(10 + 5 * vdw_correction):`), and
this was not creating a `count2` variable, and a NameError was raised,
then swallowed and hidden by the generic `except Exception`.

I now:
- replace directly the search on the next 15 lines in all cases
  (this should be safe in all cases, and shouldn't be a major performance hit)
- set `count2 = None` before, and issue a Warning if `P=` is not found
- print additional output (traceback) in the warning, in case
  an exception is raised.

The latter might need to be done elsewhere, but for now I think it's safe
and a good thing to start doing it here.