Implement parallelization options as explicit inputs.
Implements the parallelization options as a single `Dict`
given in the `parallelization` input.

If any parallelization option is specified twice (either
both in `parallization` and in `settings['CMDLINE']`, or
more than once in `settings['CMDLINE']`), an
`InputValidationError` is raised.
If the flag is given only in `settings['CMDLINE']`, an
`AiidaDeprecationWarning` is issued.
If any parallelization option is already specified in the
`settings['CMDLINE']`, a `AiidaDeprecationWarning` is issued.

The `BasePwCpInputGenerator` is given three class attributes:
- `_PARALLELIZATION_FLAGS` is a mapping `{flag_name: help}`
   of all possible flags.
- `_ENABLED_PARALLELIZATION_FLAGS` is a tuple of flag names that
  are implemented in a particular code.
- `_PARALLELIZATION_FLAG_ALIASES` is used to detect all possible
  variations on a flag name. These are taken from the QE source.

When checking for existing parallelization flags in the manually
passed cmdline parameters, these are normalized by splitting on
whitespace.
QE ignores flags that are capitalized differently, so we do not
have to normalize capitalization here.

Currently, `_ENABLED_PARALLELIZATION_FLAGS` is set only for PW
calculations. Technically it accepts the `-nimage` flag, but we
disallow it because (AFAIK) it doesn't make any sense - it just
repeats the same calculation.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>