`PwCalculation`: hotfix the `qes_200420.xsd` XML schema file
The original schema file required all van der Waals functionals to be
all capital, even though the input accepts also all lower case. A patch
was submitted to the schema:

https://github.com/QEF/qeschemas/blob/912e059ae52c076616f132367b763fa0741a3fc7/PW_CPV/qes_devel_030920.xsd

which we are already applying here. By simply correcting the existing
schema file, all calculations with output generated by that schema
should still parse fine, but the bug with respect to the functional
capitalization should now be fixed.

The existing test for the `qes_200420.xsd` test is updated to on purpose
include a functional input with lower case to make sure the parser does
not except.