`PwParser`: fix bug in raw parsing of van der Waals and stress (#550)
Previously, the flag `vdw_correction` was set in the raw pw parser only
if the string 'Carrying out vdW-DF run using the following parameters:'
was found. However, vdW-DF is not the only vdW functional, there are
many more. For example the rVV10 functional, the string was more like
'Carrying out RVV10 run...', and the capitalization might also differ.

Since the `vdw_correction` flag is used later to parse lines having the
text 'Non-local correlation energy' in it, the setting of the flag is
now based on this string instead, which should be universal.

This bug manifested with a generic 'Error while parsing stress tensor.'.
The reason is that the search for `P=` was failing since the previously
required `vdw_correction` was not found because it was not searching for
enough down the file.

This commit changes a few things in the raw parser to address this:

 * Replace directly the search on the next 15 lines in all cases, which
   should be safe in all cases, and shouldn't be a major performance hit
 * Set `count2 = None` before, and issue a Warning if `P=` is not found
 * Print additional output (traceback) in the warning, in case an
   exception is raised.