`PwBandsWorkChain`: incorrect number of bands for non-collinear spin (#570)
The workchain would compute the number of bands to use based, among
other things, on the number of spin components. This is an integer that
represents the spin polarization mode, where `npsin=1` means unpolarized
and `npsin=2` is spin-polarized. The number of bands is multiplied by
this number because the spin-polarized also happens to require twice the
number of bands.

However, there is also `npsin=4` which is used to enable non-collinear
spin. Since we were using the `nspin` parameter, which is really just
a mode integer, to be a multiplier, the non-collinear case would multiply
the number of bands by four (instead of two) causing twice as many bands
to be used as expected.