`Parsers`: remove explicit check for the retrieved folder's existence (#580)
As of `aiida-core==1.4.0`, the engine itself will check for the presence
of the retrieved folder. If it is missing, which is really just a bug in
`aiida-core` the engine will terminate the process and not call the
parser. Anyway there would be no point since there is nothing to parse.
This means that now if the parser is called, it is guaranteed that the
retrieve folder exists and so we no longer have to explicitly check for
it.

Since this is only valid from `aiida-core>=1.4` we update the dependency
requirement. We also require `aiida-core>=1.4.2` because `1.4.0` and
`1.4.1` contain a bug where the content of files in `local_copy_list`
are copied to the repository even though they shouldn't. For this plugin
this means that mostly the content of the pseudo potentials are
duplicated every time in the repository, which will bloat the repository
unnecessarily.