`ProjwfcParser`: support kind names with numerals, dashes and underscores (#591)
The regex defined in the `find_orbitals_from_statelines` parsing method
to parse the kind names, only matched alpha numeric characters, even
though the Quantum ESPRESSO spec clearly allows numbers, dashes and
underscores to be used as well. The regex is updated accordingly.

To test this, the output of the `default` test is modified manually to
use `Si1` as the kind name, which now functions as a regression test.

Co-authored-by: Kristjan Eimre <eimrek@users.noreply.github.com>

Cherry-picked from 216015807adb255e95df8a0c80a2379ea4834818