CLI: add tests for the process launch commands (#589)
The CLI had no tests whatsoever, and especially for the process launch
commands this was becoming critical, as they contain quite some code and
they are used a lot for demonstration purposes. Here we add tests for
all commands that launch a process, just by passing the required inputs.
This tests that the command can be invoked and that the code path for
the default inputs can be executed without problem.

To prevent the process from actually being submitted, the utility method
`cli.utils.launch.launch_process` is monkey patched to be a no-op. The
code of this utility should be tested once separately.