`PwCalculation`: fix bug when `hubbard_file` is specified (#596)
The `prepare_for_submission` was using the `filename` of the
`hubbard_file` input as the UUID in the `local_copy_list` which
obviously fails. In addition, the `input_file_name_hubbard_file`
classproperty of the `HpCalculation` clas was changed and adapted here
to `filename_input_hubbard_parameters`.

Note that a test cannot be added because it would have to depend on the
`aiida-quantumespresso-hp` plugin, which itself depends on
`aiida-quantumespresso`, creating a circular dependency. The test will
be added to the former to make sure this behavior works.