Remove parser information from `output_parameters` (#597)
Now that `aiida-core` adds both the AiiDA and plugin version to the
`CalcJobNode` attributes, as well as the parser name, it is no longer
necessary to add this information to the output_parameters dictionary.
This PR removes the `get_parser_info()` function that generates this
output.

The `parser_warnings` are also removed for the base/pw/ph parsers, since
they were not used here. We leave the `parser_warnings` for the neb/cp
parsers, but these should be replaced in a future PR by the use of a
logging container.