`PwBaseWorkChain`: deprecate the `pseudo_family` input
The use of a `pseudo_family`, which has to be passed in as a `Str` node
to represent the name of a `Group`, is not robust with respect to the
provenance. Since groups are mutable, rerunning a workchain that used
the `pseudo_family` to determine what pseudos to run, can lead to a
rerun using different pseudos. This input exists for historical reasons
because it made the usage of the workflows easier for users, since they
wouldn't have to build the mapping of pseudos themselves, but with the
advent of `aiida-pseudo` and the input generators, this is no longer an
issue and it is preferable to directly pass the pseudos into the
`pseudos` input namespace.

Also deprecate the functions in `utils.pseudopotential`, one of which is
not even used currently, and the other two which will be obsolete as
they are replaced by method directly integrated into the pseudo family
groups.