`PwBaseWorkChain`: restart from scratch when convergence not reached (#604)
This was already being done by the handler for relax/vc-relax
calculations where electronic convergence was not reached, but for
normal scf/nscf calculations, the restart was done from the last
calculation. It is not yet clear which approach will guarantee a higher
convergency rate, as in some cases restarting from scratch is better,
whereas in others it is detrimental, but at least the approaches should
be the same. The handler for `ERROR_ELECTRONIC_CONVERGENCE_NOT_REACHED`
is therefore updated to also restart from scratch, as the docstring
already said it would do.