Dependencies: update the `aiida-pseudo` requirements (#609)
The new required version of `aiida-pseudo` allows multiple cutoff sets
to be defined on a pseudopotential family at different levels of
stringency. It also changes the CLI option to pass a pseudopotential
family to `-F`.

The `UpfData` class was also updated to now require the `z_valence`
value to be parsed in addition to the element, so the test fixtures are
updated to include that in the generated pseudo nodes.

The tests are adapted and a new fixture `sssp` is added, which for the
whole session creates an instance of `SsspFamily`. Since this is a
costly operation, it should only be done once and all tests should
reuse it. This means that tests can no longer blindly use the fixture
`clear_database_before_tests`, which anyway was not useful in terms of
efficiency.