`PwParser`: ignore stress threshold if `CELL.cell_dofree != 'all'`
If the `cell_dofree` is anything other than the default `all` where no
degrees of freedom of the cell are constrained, the stress convergence
threshold needs to be ignored. The reason is that often the stress and
therefore pressure are far from a minimum because the cell if articially
constrained.