`PwRelaxWorkChain`: replace `relaxation_scheme` with `relax_type`
The `relaxation_scheme` was used to communicate the type of relaxation
to be performed and accepted either `relax` or `vc-relax`, directly
modeled after the relax options that could be passed to the
`CONTROL.calculation` input parameter.

There are, however, many other relax modes possible, such as those
constraining parts of the degree of freedom of the cell. To support
these a new input is introduced `relax_type` that replaces the old
`relaxation_scheme` that is deprecated. The `relax_type` can take values
as specified by the `RelaxType` enum.