`PwRelaxWorkChain`: do not attach output structure for `RelaxType.NONE` (#616)
This relax type was mostly added for the common workflows project where
they want to leverage the common relax workflows to also simply provide
a single point calculation to just compute the total energy and forces
for a given structure. In this case, the structure won't change and so
attaching it as an output makes no sense.