`PwCalculation`: add explicit `parallelization` input port (#554)
This new input port accepts a dictionary that allows one to specify the
parallelization flags that `pw.x` accepts. This supersedes the old method
of defining the flags in the `settings `input node. The latter is
deprecated but still takes precedence if the same flag is specified
through both input ports.