Fix projwfc parser for >=1000 projections
Fix a bug in the projwfc parser when >= 1000 projections are
used, and there is one fewer whitespace in the energy line.

Instead of splitting by whitespace alone, we now split at the
equals sign with surrounding whitespaces.

In the tests, we use a bit of a hack to test for this; replacing
a '1' with '0001' to force the old behavior to break without
having to add >=1000 projections.