`ProjwfcParser`: fix bug in case of for more than 1000 orbitals (#624)
The parser was relying on spaces being present between the parentheses
that contain the index of the orbital. Starting at four characters, i.e.
starting from orbital 1000, no more spaces are printed for padding and
so the splitting failed.

Instead of splitting by whitespace alone, we now split at the equals
sign with surrounding whitespaces. In the tests, we use a bit of a hack
to test for this; replacing a `1` with `0001` to force the old behavior
to break without having to add more than a 1000 projections.