`Protocol`: Switch to initial_magnetic_moments
Replace the `starting_magnetization` input argument of the
`get_builder_from_protocol` method of the `PwBaseWorkChain` by
`initial_magnetic_moments`.

Allowing the user to define the initial magnetic moments in Bohr
magnetons instead of the starting_magnetization has several advantages:

* It's easier to think of an initial guess for the magnetic moment than
the starting_magnetization based on e.g. the oxidition state of the
element.
* The magnetization output of pw.x also is expressed in Bohr magnetons.
* Since deriving the `starting_magnetization` from the magnetic moments
requires the `z_valence`, it requires knowledge about which
pseudopotentials are being used for the calculation. This makes it a
more difficult quantity to work with than the magnetic moments when
using the `get_builder_from_protocol` method in other projects.