`Devops`: Fix MANIFEST.in for .xsd files
When the XML schemas were updated for the cp.x code (b926c72), the .xsd
files were moved, which caused them to be no longer included in the
source build.

Here we update the MANIFEST,in file to recursively look for .xsd files
in the aiida_quantumespresso/parsers/parse_xml folder, do the current
files are properly added to the build and future additions of schemas
for other codes should be included as well.