Protocols: fix mutation of overrides (#650)
Currently the `overrides` input dictionary provided to the
`get_builder_from_protocol` method can be mutated to also include the
defaults from the protocol. Here we make a copy of the `right`
dictionary in the `recursive_merge` function, to prevent the key/value
pairs of the `left` dictionary to be added to the `right` one. This
prevents the mutation of the `overrides` argument.