`RelaxType`: Rename `ATOMS` to `POSITIONS`
Renames the `RelaxType.ATOMS` to `RelaxType.POSITIONS`, since it is the
positions of the atoms that are being relaxed. Similarly,
`ATOMS_VOLUME`, `ATOMS_SHAPE` and `ATOMS_CELL` are renamed to
`POSITIONS_VOLUME`, `POSITIONS_SHAPE` and `POSITIONS_CELL`,
respectively.

This change is also necessary to accommodate a similar change in
`aiida-common-workflows`, see:

https://github.com/aiidateam/aiida-common-workflows/pull/146