`BasePwCpInputGenerator`: Improve parameters input validation
First working example of how we can add validation of the input
parameters of the Quantum ESPRESSO calculation jobs by converting the
INPUT-{}.XML files that are generated for the QE documentation into JSON
schemas that are then used in the validator of the `parameters` input.

The `convert_to_json_schema()` function in the `calculations.schemas.utils.py`
module is used to generate the `calculations.schemas.pw.parameters.yaml`
file, based on the `calculations.helpers.INPUT_PW-6.4.xml` file. This
yaml file is then loaded in the validator for the `parameters` input and
used to check if the parameters adhere to the prescriptions of the JSON
schema.