Protocols: Add support for `CutoffsPseudoPotentialFamily`
In the `aiida-pseudo` release v1.6.1, a generic pseudo family class was added
which has the features of the `RecommendedCutoffsMixin` class. This
allows the user to install pseudo families other than the supported SSSP
and pseudo-dojo configurations and set recommended cutoffs for specified
stringencies.

Here we add support for the `CutoffsPseudoPotentialFamily` class to the
`get_builder_from_protocol` method of the `PwBaseWorkChain`. Since these
pseudo families aren't configured with recommended cutoffs by default,
we catch any ValueErrors raised by the `get_recommended_cutoffs` method
and add reraise them with a message that indicates that indicates that
the failure is related to obtaining the recommended cutoffs.