`PwBandStructureWorkChain`: Deprecate and fix (#688)
The `PwBandStructureWorkChain` is a legacy version of a band structure work
chain with a pre-defined protocol, mostly used for previous tutorials.
However, now that the `PwBandsWorkChain` has been equipped with a protocol
through the `get_builder_from_protocol()` method, it no longer serves a
purpose and should be deprecated.

However, there were still two issues with running the work chain that
needed to be fixed, since users were still running it and it is still
supported until v4.0.0:

* The `get_pseudos_from_dict()` method, used by the work chain, only
accepted `UpfData` from the `aiida-pseudo` package. Since the work chain
was often run with legacy `UpfData`, this raised errors for several
users. This is fixed by also accepting the legacy `UpfData` here.
* There was a small bug in the `PwRelaxWorkChain` when using the
`relaxation_scheme` input (which is also deprecated). Basically, at some
point in the work chain, the `base` input port was checked for the
`pw.parameters.CONTROL.calculation` value, but in case the user doesn't
specify this input (which passes validation when the `relaxation_scheme`
is provided), this would raise a `KeyError` when running the work chain.
This is fixed by obtaining the `calculation` value from the context
instead.