`ProtocolMixin`: adapt loading of protocol file
The `load_protocol_file` method in the protocol utils.py module uses
the path of the module file to obtain the path to the .yaml file that
described the protocols. However, this fails in case the `ProtocolMixin`
class is used by other packages, since the protocol files of the
package's work chains will be stored in that package.

Here we adapt the `ProtocolMixin` class as follows:

 * Add a `get_protocol_filepath()` class method that needs to be
   implemented by the work chain classes that use the `ProtocolMixin`.

 * Move the `load_protocol_file` method into the `ProtocolMixin` class as
   a hidden method that relies on the `get_protocol_filepath` method.

Moreover, the `Path` to the hardcoded protocols that ship with the package
are now also obtained using `importlib_resources`. It is added as a
dependency as this was backported from Python 3.9.