`ProtocolMixin`: allow `overrides` to be passed as a `Path`
Currently the user can only adapt the protocol by specifying `overrides`
as a dictionary, or adapting the builder afterwards.

Here the `get_protocol_inputs` method is made more flexible by accepting
a `pathlib.Path` for the `overrides` input. This allows the user to
define protocol overrides in a `.yaml` file.