Docs: change theme and overhaul structure
The documentation is currently quite outdated, and the main page was
quite an overwhelming amount of information, with very little guidance
on where to begin.

Here we switch to the `sphinx-book-theme` used by several other AiiDA
projects and clean up the layout somewhat. We also add:

* The CLI reference
* An installation section on the main page
* A Getting started section on the main page that points to the AiiDA
Quantum ESPRESSO tutorial on the AiiDA tutorials readthedocs.
* A warning on the main page that indicates that some of the material
may be outdated.