`PwBaseWorkChain`: add handler for intentional non-convergence (#695)
In certain cases, a user will set inputs such that the SCF will not
converge, for example, by setting `scf_must_converge` to false, or by
setting `electron_maxstep` to zero. This can be useful to initialize a
charge density of a system and the workchain should not restart or fail
in this case.

The `PwParser` now returns `WARNING_ELECTRONIC_CONVERGENCE_NOT_REACHED`
instead of `ERROR_ELECTRONIC_CONVERGENCE_NOT_REACHED` if any of these
parameters are set as described. The `PwBaseWorkChain` has a new handler
for this new exit code to consider the calculation as successful and
finish the workchain.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>