Dependencies: temporarily set upper limit for `psycopg2` (#707)
A recent release of `psycopg2 v2.9` has broken (at least) the
continuous integration tests suite. This was already fixed in
`aiida-core v1.6.x` but not (yet?) backported to the last version
that support Python 3.6 which is still supported by this plugin.

This pin can be removed once a higher version of `aiida-core` is
supported or once `psycopg2` has released a fix for these problems.