Tests: fix immigrate tests after change in `aiida-core` (#709)
The nightly build started failing because a recent change in
`aiida-core`. The `CalcJob` now has a new namespace in the metadata
options called `stash`. This means that the options of any process builder
will now include this namespace even if nothing is specified.

This change was causing the immigrate tests to fail since the inputs
options do not have this namespace and so the comparison was failing.
Since not all versions of `aiida-core` have this change, such as the
Python 3.6 compatible versions, we simply remove the namespace from the
builder if it is present before doing the comparison with the inputs.