`PwBaseWorkChain`: improve restart from `parent_folder`
Remove some of the logic in the `PwBaseWorkChain` regarding restarting
from a previous calculation using a `RemoteData` provided to the
`pw.parent_folder` input.

The current logic expected the `RemoteData` to have a `PwCalculation`
creator, which is not always the case. Moreover, the `restart_mode`
chosen by the user is overriden, which means that e.g. restarting from
_only_ the charge density with `startingpot` is not possible.