`PwParser`: add support for QE v6.8 with new XML schema file (#717)
QE v6.8 has been released but the XML output it produces follows a new
schema. The schema is added such that the `PwParser` can parse the XML
content.

Note that due to a bug in the output writing code, the output actually
does not respect the schema and so parsing will fail with an error. The
bug has been fixed in this commit:

    https://gitlab.com/QEF/q-e/-/merge_requests/1531

However, the fix is not expected to be released in a patch version but
will rather ship with the next major release. Therefore, we add an
exception in our own parser to ignore the duplicated value in the XML
output if it is produced by QE v6.8. In this way, the database will not
be spammed with unnecessary warning logs that have no further influence
on the quality of the results.