`PwParser`: Fix magnetic moments for QE v6.8
Due to a change in the format of the output of the magnetic moments per
site, the parsing of spin-polarised `pw.x` calculations was broken.

Here we add a few lines of parsing that fixes the issue for v6.8.
Basically, if the line doesn't contain `atom:` with a colon, the parser
now also checks if `atom` is there, and parses the atomic charges and
magnetic moments correctly in this case.