`PwBaseWorkChain`: make magnetism from `overrides` absolute
When using the `get_builder_from_protocol()` method from the
`PwBaseWorkChain`, setting `spin_type` to `COLLINEAR` while not
specifying the `initial_magnetic_moments` will result in a set of
default magnetic moments being used, based on the `magnetization.yaml`
file. This means that even when `SYSTEM.starting_magnetization` is
specified in the `overrides`, these are simply ignored in favor of the
defaults.

Here we make the `overrides` _absolute_, i.e. the
`starting_magnetization` is only set to the default in case it is not
specified in the `overrides`. This is once again based on the principle
that we should avoid quietly overriding inputs that the user explicitly
specified.