`README.md`: cleanup the syntax of compatibility matrix
The shields for the `aiida-core` compatibility use a custom AiiDA logo
passed in the URL parameters as a base-64 encoded string. This long
string makes the table difficult to read. The links are now defined
using reference links allowing the long links to be placed at the bottom
of the document. Note that there is still a lot repetition as for each
link the image string needs to be repeated, but infortunately, basic
markdown does not seem to support variable substitution.