`README.md`: add shields for Quantum ESPRESSO compatibility (#739)
Encode the Quantum ESPRESSO logo for the shield as a base64 encoded
string. The image used replaces the white background of the original
logo with #5d5d5d which is the background used by the shield pills. We
cannot use transparency, because that will make the image of the png too
large, since it requires saving a separate transparency channel, and
img.shields.io will return a 400 Bad Request.