`PwBaseWorkChain`: handle diagonalization issues
Add parsing for an error related to the diagonalization, and add a
corresponding exit code to the `PwCalculation`.

Add an error handler to deal with both this exit code as well as the
`ERROR_COMPUTING_CHOLESKY` exit code. In both cases, switching to using
conjugate gradient for the diagonalization seems to be a quite
successful approach for recovering from these failures. If the
calculation was already running conjugate gradient, the
`PwBaseWorkChain` fails with a known unrecoverable failure.