`PwCalculation`: adjust `parent_folder` validation
Currenty, work chains that wrap the `PwCalculation` for an `nscf` or
`bands` calculation cannot be used in case they generate the
`parent_folder` input inside the work chain. This is because they fail
to pass validation since now the `PwCalculation` expects a
`parent_folder` input for these calculations.

Here we split up the validation of the top-level inputs of the
`PwCalculation` into two parts:

* `validate_inputs_base` checks _only_ the input parameters for `nscf`
  and `bands`calculations.
* `validate_inputs` checks the presence of the `parent_folder` for `nscf`
  and `bands` calculations and subsequently calls the
  `validate_inputs_base` validator.

The default validation for the `PwCalculation` top-level inputs is
`validate_inputs`, so both validation methods are called. Work chains
that wrap the `PwCalculation` and generate the `parent_folder` inside
the work chain can set the validator for the `PwCalculation` port to
`PwCalculation.validate_inputs_base`, as is now done for the
`PdosWorkChain`.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>